requirements.txt
streamlit
pandas
firebird-driver
plotly